Tom Bartol:
  Modified struct counter again.  Hit counts are doubles (there could be
  more than 4 billion hits counted during a time step) but contents counts
  are ints (you can't have more than 4 billion molecules in a simulation).
  Region counting is now working 100% (plus or minus)...

+++ src/mcell_structs.h
@@ -729,10 +729,10 @@ struct counter
   struct counter *next;
   struct region *reg_type;
   struct species *mol_type;
-  int front_hits;
-  int back_hits;
-  int front_to_back;
-  int back_to_front;
+  double front_hits;
+  double back_hits;
+  double front_to_back;
+  double back_to_front;
   int n_inside;
 };
